# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Expense Tracker API
  description: API для отслеживания расходов с аналитикой
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
    description: Локальный сервер разработки


paths:
  /:
    get:
      summary: Корневой endpoint
      description: Возвращает приветственное сообщение
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: string
                example: "Добро пожаловать в Expense Tracker API"

  /v1/healthcheck:
    get:
      summary: Проверка состояния сервиса
      description: Возвращает статус работы сервиса
      responses:
        '200':
          description: Сервис работает
          content:
            application/json:
              schema:
                type: string
                example: "OK"

  /v1/help:
    get:
      summary: Список доступных команд
      description: Возвращает список всех доступных эндпоинтов API
      responses:
        '200':
          description: Список эндпоинтов
          content:
            application/json:
              schema:
                type: string
                example: "Доступные команды: /expenses/add, /expenses/most_expensive_category, /expenses/most_expensive_purchase"

  /v1/expenses:
    post:
      summary: Добавить новый расход
      description: Добавляет запись о новом расходе в систему
      requestBody:
        required: true
        content:
          application/json:
            example:
              expense: string,
              category: string
              amount: string
              date: string
      responses:
        '200':
          description: Расход успешно добавлен
          content:
            application/json:
              schema:
                type: string
                example: "Расход добавлен"
        '422':
          description: Ошибка валидации
          content:
            application/json:
              example:
                detail:
                    loc: [string, 0]
                    msg: string
                    type: string


  /v1/expenses/{date}:
    get:
      summary: Самая затратная категория
      description: Возвращает категорию с наибольшими расходами за указанную дату
      parameters:
        - name: date
          in: path
          required: true
          description: Дата в формате YYYY-MM-DD
          schema:
            type: string
            example: "2023-11-15"
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: string
                example: "Продукты"
        '422':
          description: Ошибка валидации
          content:
            application/json:
              example:
                detail:
                    loc: [string, 0]
                    msg: string
                    type: string

  /v1/expenses/{category}/{date}:
    get:
      summary: Самая дорогая покупка
      description: Возвращает самую дорогую покупку в указанной категории за дату
      parameters:
        - name: category
          in: path
          required: true
          description: Название категории
          schema:
            type: string
            example: "Продукты"
        - name: date
          in: path
          required: true
          description: Дата в формате YYYY-MM-DD
          schema:
            type: string
            example: "2023-11-15"
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                type: string
                example: Молоко - 150 руб.
        '422':
          description: Ошибка валидации
          content:
            application/json:
              example:
                detail:
                    loc: [string, 0]
                    msg: string
                    type: string